Inventory Service implementation

+++ src/main/java/com/inventory/service/controller/InventoryServiceController.java
@@ -1,16 +1,17 @@
 package com.inventory.service.controller;
 
-import com.inventory.service.exceptions.BrandNotFoundException;
-import com.inventory.service.exceptions.CategoryNotFoundException;
-import com.inventory.service.model.Brand;
-import com.inventory.service.model.Category;
+import com.inventory.service.exceptions.*;
+import com.inventory.service.model.*;
 import com.inventory.service.service.*;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 public class InventoryServiceController {
@@ -70,6 +71,10 @@ public class InventoryServiceController {
 
     @DeleteMapping(value = "/inventory/brand/delete/{brandId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<String> deleteBrand(@PathVariable Long brandId) {
+        List<Product> productsByBrand = productService.findProductsByBrand(brandId);
+        if (!CollectionUtils.isEmpty(productsByBrand)) {
+            throw new BrandNotEmptyException("Can't delete brand with id " + brandId + ", brand has products.");
+        }
         try {
             brandService.deleteBranch(brandId);
         }
@@ -97,11 +102,46 @@ public class InventoryServiceController {
 
     @DeleteMapping(value = "/inventory/category/delete/{categoryId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<String> deleteCategory(@PathVariable Long categoryId) {
+        List<Product> productsByCategory = productService.findProductsByCategory(categoryId);
+        if (!CollectionUtils.isEmpty(productsByCategory)) {
+            throw new CategoryNotEmptyException("Can't delete category with id " + categoryId + ", category has products.");
+        }
         try {
             categoryService.deleteCategory(categoryId);
         }
         catch (EmptyResultDataAccessException e) {
             throw new CategoryNotFoundException("Not found category with id: " + categoryId);
+        }
+        return ResponseEntity.ok().build();
+    }
+
+    /***********/
+    /** STORE **/
+    /***********/
+
+    @PostMapping(value = "/inventory/store/add", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Store> addStore(@RequestBody Store store) {
+        Store savedStore = storeService.saveOrUpdateStore(store);
+        return ResponseEntity.status(HttpStatus.CREATED).body(store);
+    }
+
+    @PutMapping(value = "/inventory/store/update", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Store> updateStore(@RequestBody Store store) {
+        Store savedStore = storeService.saveOrUpdateStore(store);
+        return ResponseEntity.ok().body(store);
+    }
+
+    @DeleteMapping(value = "/inventory/store/delete/{storeId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> deleteStore(@PathVariable Long storeId) {
+        List<Item> itemsByStore = itemService.findItemsByStore(storeId);
+        if (!CollectionUtils.isEmpty(itemsByStore)) {
+            throw new StoreNotEmptyException("Can't delete store with id " + storeId + ", store has products.");
+        }
+        try {
+            storeService.deleteStore(storeId);
+        }
+        catch (EmptyResultDataAccessException e) {
+            throw new StoreNotFoundException("Not found store with id: " + storeId);
         }
         return ResponseEntity.ok().build();
     }
